fixing bug by creating a new file for the component
