+++ client/src/pages/Booking.tsx
@@ -1,4 +1,4 @@
-import { ProtectedCreateBookingSection } from "components/composite/BookingCreation/BookingCreation"
+import { ProtectedCreateBookingSection } from "components/composite/BookingCreation/ProtectedCreateBookingSection"
 import { Route, Routes } from "react-router-dom"
 import BookingSuccess from "./BookingSuccess"
 import { Footer } from "components/generic/Footer/Footer"
